Data Module: Add support for action creators as generators
