+++ packages/data/src/index.js
@@ -20,6 +20,7 @@ import isShallowEqual from '@wordpress/is-shallow-equal';
  * Internal dependencies
  */
 import registerDataStore from './store';
+import createStoreRuntime from './runtime';
 
 export { loadAndPersist, withRehydration, withRehydratation } from './persist';
 
@@ -27,6 +28,7 @@ export { loadAndPersist, withRehydration, withRehydratation } from './persist';
  * Module constants
  */
 const stores = {};
+const runtimes = {};
 const selectors = {};
 const actions = {};
 let listeners = [];
@@ -97,6 +99,9 @@ export function registerReducer( reducerKey, reducer ) {
 		}
 	} );
 
+	// Create the actions runtime
+	runtimes[ reducerKey ] = createStoreRuntime( store );
+
 	return store;
 }
 
@@ -147,6 +152,7 @@ export function registerResolvers( reducerKey, newResolvers ) {
 		}
 
 		const store = stores[ reducerKey ];
+		const runtime = runtimes[ reducerKey ];
 
 		// Normalize resolver shape to object.
 		let resolver = newResolvers[ selectorName ];
@@ -165,20 +171,9 @@ export function registerResolvers( reducerKey, newResolvers ) {
 			// state, it would not be otherwise provided to fulfill.
 			const state = store.getState();
 
-			let fulfillment = resolver.fulfill( state, ...args );
+			const fulfillment = resolver.fulfill( state, ...args );
 
-			// Attempt to normalize fulfillment as async iterable.
-			fulfillment = toAsyncIterable( fulfillment );
-			if ( ! isAsyncIterable( fulfillment ) ) {
-				return;
-			}
-
-			for await ( const maybeAction of fulfillment ) {
-				// Dispatch if it quacks like an action.
-				if ( isActionLike( maybeAction ) ) {
-					store.dispatch( maybeAction );
-				}
-			}
+			await runtime( fulfillment );
 
 			finishResolution( reducerKey, selectorName, args );
 		}
@@ -212,8 +207,8 @@ export function registerResolvers( reducerKey, newResolvers ) {
  * @param {Object} newActions   Actions to register.
  */
 export function registerActions( reducerKey, newActions ) {
-	const store = stores[ reducerKey ];
-	const createBoundAction = ( action ) => ( ...args ) => store.dispatch( action( ...args ) );
+	const runtime = runtimes[ reducerKey ];
+	const createBoundAction = ( action ) => ( ...args ) => runtime( action( ...args ) );
 	actions[ reducerKey ] = mapValues( newActions, createBoundAction );
 }
 
@@ -401,77 +396,5 @@ export const withDispatch = ( mapDispatchToProps ) => createHigherOrderComponent
 	] ),
 	'withDispatch'
 );
-
-/**
- * Returns true if the given argument appears to be a dispatchable action.
- *
- * @param {*} action Object to test.
- *
- * @return {boolean} Whether object is action-like.
- */
-export function isActionLike( action ) {
-	return (
-		!! action &&
-		typeof action.type === 'string'
-	);
-}
-
-/**
- * Returns true if the given object is an async iterable, or false otherwise.
- *
- * @param {*} object Object to test.
- *
- * @return {boolean} Whether object is an async iterable.
- */
-export function isAsyncIterable( object ) {
-	return (
-		!! object &&
-		typeof object[ Symbol.asyncIterator ] === 'function'
-	);
-}
-
-/**
- * Returns true if the given object is iterable, or false otherwise.
- *
- * @param {*} object Object to test.
- *
- * @return {boolean} Whether object is iterable.
- */
-export function isIterable( object ) {
-	return (
-		!! object &&
-		typeof object[ Symbol.iterator ] === 'function'
-	);
-}
-
-/**
- * Normalizes the given object argument to an async iterable, asynchronously
- * yielding on a singular or array of generator yields or promise resolution.
- *
- * @param {*} object Object to normalize.
- *
- * @return {AsyncGenerator} Async iterable actions.
- */
-export function toAsyncIterable( object ) {
-	if ( isAsyncIterable( object ) ) {
-		return object;
-	}
-
-	return ( async function* () {
-		// Normalize as iterable...
-		if ( ! isIterable( object ) ) {
-			object = [ object ];
-		}
-
-		for ( let maybeAction of object ) {
-			// ...of Promises.
-			if ( ! ( maybeAction instanceof Promise ) ) {
-				maybeAction = Promise.resolve( maybeAction );
-			}
-
-			yield await maybeAction;
-		}
-	}() );
-}
 
 registerDataStore();
